requirements.txt
pandas>=1.5.0
numpy>=1.21.0
websocket-client>=1.3.0
python-binance>=1.0.16
matplotlib>=3.5.0
plotly>=5.10.0
python-dotenv>=0.20.0
ccxt>=2.0.0
ta>=0.10.0
